Create new library entry point
- Create `SummitAST` as entry point for Summit AST library
- Move library functions from `SummitTool` to `SummitAST`
- Separate build targets
