+++ src/main/java/com/google/summit/SummitTool.kt
@@ -16,102 +16,21 @@
 
 package com.google.summit
 
-import com.google.common.base.Ascii
 import com.google.common.flogger.FluentLogger
-import com.google.summit.translation.Translate
-import com.nawforce.apexparser.ApexLexer
-import com.nawforce.apexparser.ApexParser
-import com.nawforce.apexparser.CaseInsensitiveInputStream
 import java.io.IOException
 import java.nio.file.Files
 import java.nio.file.Path
 import java.nio.file.Paths
 import java.util.stream.Stream
-import org.antlr.v4.runtime.BaseErrorListener
-import org.antlr.v4.runtime.CharStreams
-import org.antlr.v4.runtime.CommonTokenStream
-import org.antlr.v4.runtime.ParserRuleContext
-import org.antlr.v4.runtime.RecognitionException
-import org.antlr.v4.runtime.Recognizer
 
-/** This is a simple command line tool to parse and translate Apex source files.
-  *
-  * Pass arguments that are file paths or directories to search. All files with
-  * the *.cls extension will be read.
-  */
+/**
+ * This is a simple command line tool to parse and translate Apex source files.
+ *
+ * Pass arguments that are file paths or directories to search. All files with the *.cls extension
+ * will be read.
+ */
 object SummitTool {
-  val logger = FluentLogger.forEnclosingClass()
-
-  /** Listener for syntax errors that keeps a total count. */
-  class SyntaxErrorListener : BaseErrorListener() {
-    public var numErrors = 0
-
-    override fun syntaxError(
-      recognizer: Recognizer<*, *>,
-      offendingSymbol: Any?,
-      line: Int,
-      charPositionInLine: Int,
-      msg: String,
-      e: RecognitionException?
-    ) {
-      this.numErrors += 1
-      logger.atInfo().log("Syntax error at %d:%d: %s", line, charPositionInLine, msg)
-    }
-  }
-
-  /** Parses and translates a single Apex source file and returns success. */
-  private fun parseAndTranslate(path: Path): Boolean {
-    // Apex is a case-insensitive language and the grammar is
-    // defined to operate on fully lower-cased inputs.
-    val lowerCasedStream = CaseInsensitiveInputStream(CharStreams.fromPath(path))
-    val lexer = ApexLexer(lowerCasedStream)
-    val tokens = CommonTokenStream(lexer)
-    val parser = ApexParser(tokens)
-
-    val errorCounter = SyntaxErrorListener()
-    lexer.addErrorListener(errorCounter)
-    parser.addErrorListener(errorCounter)
-
-    // Do parse as complete compilation unit
-    val tree = when {
-      isApexClassFile(path) -> parser.compilationUnit()
-      isApexTriggerFile(path) -> parser.triggerUnit()
-      else -> throw IllegalArgumentException("Unexpected file type")
-    }
-
-    if (errorCounter.numErrors > 0) {
-      logger.atWarning().log("Failed to parse %s", path)
-      return false // failure
-    }
-
-    try {
-      val translator = Translate(path.toString(), tokens)
-      translator.translate(tree)
-    } catch (e: Translate.TranslationException) {
-      logger.atWarning().withCause(e).log("Failed to translate %s", path)
-      return false // failure
-    }
-
-    return true
-  }
-
-  /**
-   * Returns true if the path is an Apex class file. These are regular files with the
-   * (case-insensitive) suffix ".cls".
-   */
-  private fun isApexClassFile(path: Path): Boolean =
-    Files.isRegularFile(path) && Ascii.toLowerCase(path.toString()).endsWith(".cls")
-
-  /**
-   * Returns true if the path is an Apex trigger file. These are regular files with the
-   * (case-insensitive) suffix ".trigger".
-   */
-  private fun isApexTriggerFile(path: Path): Boolean =
-    Files.isRegularFile(path) && Ascii.toLowerCase(path.toString()).endsWith(".trigger")
-
-  /** Returns true if the path is an Apex source file: either a class or a trigger. */
-  private fun isApexSourceFile(path: Path): Boolean =
-    isApexClassFile(path) || isApexTriggerFile(path)
+  private val logger = FluentLogger.forEnclosingClass()
 
   @JvmStatic
   fun main(args: Array<String>) {
@@ -125,12 +44,17 @@ object SummitTool {
 
       try {
         val stream: Stream<Path> =
-          Files.find(Paths.get(arg), Integer.MAX_VALUE, { path: Path, _ -> isApexSourceFile(path) })
-        stream.forEach({ path: Path ->
+          Files.find(
+            Paths.get(arg),
+            Integer.MAX_VALUE,
+            { path, _ -> SummitAST.isApexSourceFile(path) }
+          )
+
+        stream.forEach { path ->
           numFiles++
-          val success = parseAndTranslate(path)
+          val success = SummitAST.parseAndTranslate(path)
           numFailures += if (success) 0 else 1
-        })
+        }
       } catch (e: IOException) {
         logger.atWarning().withCause(e).log("Invalid path %s", arg)
       }
